ST SPEAr: Correcting CLCD clk recalc routine
CLCD clk_recalc routines in plat-spear/clock.c are incorrect. This patch
fixes mistakes in those routines and changes default tables values in
mach-spear13xx/clock.c file for clcd.

+++ arch/arm/plat-spear/include/plat/clock.h
@@ -224,7 +224,7 @@ struct clcd_clk_config {
 
 /* clcd clk rate config structure */
 struct clcd_rate_tbl {
-	u16 div;
+	u32 div;
 };
 
 /* ddr min, max clk rate table */
